requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
streamlit==1.49.1
sentence-transformers==5.1.0
transformers==4.56.1
tokenizers==0.22.0
torch==2.8.0
scikit-learn==1.7.2
huggingface-hub==0.34.4
tqdm==4.67.1
numpy==2.3.3
pandas==2.3.2
milvus-lite==2.5.1
pymilvus==2.6.1
pymupdf==1.26.4
pillow==11.3.0
omegaconf==2.3.0
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
typing-inspection==0.4.1
typing-extensions==4.15.0
requests==2.32.5
gitpython==3.1.45
gitdb==4.0.12
ollama==0.5.4
llama-cpp-python==0.3.16
fastapi-responses==0.2.1
pydantic==2.11.9
pydantic-core==2.33.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
ipykernel==6.30.1
ipython==9.5.0
matplotlib-inline==0.1.7
importlib-metadata==8.7.0
pip==25.2
